requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
ollama>=0.6.1
flask>=3.1.3